made some new goals

+++ def.h
@@ -3,7 +3,7 @@
 
 typedef struct node
 {
-  int val;
+  float val;
   struct node *U,*D,*L,*R; //up down left right will not be used in the demo
 }node;
 //double LL carring doubly LL(node)
